+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v1.0.1
 
 * Index close allows active cursors now. They will behave the same as if they were created
   after the index was closed.
+* Added ability to rename an index.
 
 v1.0.0
 ------